Improved error handling. Removed unnecessary checks

+++ include/core/token/token.hpp
@@ -16,11 +16,11 @@ enum class token_type : uint8_t;
 struct token {
     token_type type_{};
     std::string_view lexeme_;
-    size_t line_{};
-    size_t column_{};
+    uint32_t line_{};
+    uint32_t column_{};
 
     token() = default;
-    token(token_type type, std::string_view lex, size_t line, size_t column);
+    token(token_type type, std::string_view lex, uint32_t line, uint32_t column);
 
     bool is_keyword() const noexcept;
     bool is_double_literal() const noexcept;
